database: Move postgresql to common. Then delete database extensions.
