+++ core/src/db/mod.rs
@@ -1,5 +1,3 @@
 mod common;
 pub mod operator;
-mod postgresql;
-mod text;
 pub mod verifier;
