+++ core/src/db/verifier.rs
@@ -1,5 +1,7 @@
 use std::ops::{Deref, DerefMut};
 
+use crate::config::BridgeConfig;
+
 use super::common::Database;
 
 #[derive(Debug, Clone)]
@@ -8,9 +10,9 @@ pub struct VerifierMockDB {
 }
 
 impl VerifierMockDB {
-    pub fn new(db_file_path: String) -> Self {
+    pub async fn new(config: BridgeConfig) -> Self {
         Self {
-            common_db: Database::new(db_file_path),
+            common_db: Database::new(config).await.unwrap(),
         }
     }
 }
